chore(test): migrate jest setup to ts
